chore(structure): reorganize folders following scalable Next.js architecture
