+++ project/app/(dashboard)/dashboard/page.tsx
@@ -1,5 +1,5 @@
 import { TrendingUp, Users, CheckCircle, Clock, Plus } from "lucide-react"
-import { DashboardLayout } from "@/components/dashboard-layout"
+import { DashboardLayout } from "@/components/features/dashboard/dashboard-layout"
 
 export default function DashboardPage() {
   return (
