+++ project/components/shared/footer.tsx
